chore: move more files into subfolders
