+++ src/health-check/is-health-check-event.ts
@@ -1,5 +1,5 @@
 import type { DecodeEventLogReturnType } from "viem";
-import type SortedOraclesABI from "../sorted-oracles-abi.js";
+import type SortedOraclesABI from "../abis/sorted-oracles.abi.js";
 import type { HealthCheckEvent } from "../types.js";
 
 export default function isHealthCheckEvent(
