+++ src/proposal-queued/is-proposal-queued-event.ts
